test connect and login

+++ Sources/libetpan/data-types/mailstream_socket.c
@@ -296,10 +296,10 @@ static ssize_t mailstream_low_socket_read(mailstream_low * s,
   }
   
   if (socket_data->use_read) {
-    r = read(socket_data->fd, buf, count);
+    r = (int)read(socket_data->fd, buf, count);
   }
   else {
-    r = recv(socket_data->fd, buf, count, 0);
+    r = (int)recv(socket_data->fd, buf, count, 0);
 #ifdef WIN32
     if (SOCKET_ERROR == r) {
       if (WSAEWOULDBLOCK == WSAGetLastError()) {
@@ -411,7 +411,7 @@ static ssize_t mailstream_low_socket_write(mailstream_low * s,
       return 0;
   }
   
-  r = send(socket_data->fd, buf, count, 0);
+  r = (int)send(socket_data->fd, buf, count, 0);
 
 #ifdef WIN32
   if (SOCKET_ERROR == r) {
